move medium model to shared folder
